+++ lib/src/clients/agent-service/models/AgentRes.ts
@@ -1,4 +1,5 @@
-import type { Failed } from '../../../models'
+import type { ComponentId, Failed, Unhandled, LocationServiceError } from '../../../models'
+import type { AkkaLocation } from '../../location'
 
 /**
  *@category Agent Service
@@ -12,6 +13,26 @@ export type Spawned = {
 export type Killed = {
   _type: 'Killed'
 }
+/**
+ *@category Agent Service
+ */
+export type SequenceComponentStatus = {
+  seqCompId: ComponentId
+  sequencerLocation: AkkaLocation[]
+}
+/**
+ *@category Agent Service
+ */
+export type AgentStatus = {
+  agentId: ComponentId
+  seqCompsStatus: SequenceComponentStatus[]
+}
+
+export type AgentStatusSuccess = {
+  _type: 'Success'
+  agentStatus: AgentStatus[]
+  seqCompsWithoutAgent: SequenceComponentStatus[]
+}
 
 /**
  * @category Agent Service
@@ -23,3 +44,9 @@ export type SpawnResponse = Spawned | Failed
  * @category Response ADT
  */
 export type KillResponse = Killed | Failed
+
+/**
+ * @category Agent Service
+ * @category Response ADT
+ */
+export type AgentStatusResponse = Unhandled | LocationServiceError | AgentStatusSuccess