+++ lib/src/decoders/AgentDecoders.ts
@@ -1,7 +1,18 @@
 import * as D from 'io-ts/lib/Decoder'
-import type { Killed, KillResponse, Spawned, SpawnResponse } from '../clients/agent-service'
-import { FailedD } from './CommonDecoders'
+import type {
+  AgentStatus,
+  AgentStatusResponse,
+  AgentStatusSuccess,
+  Killed,
+  KillResponse,
+  SequenceComponentStatus,
+  Spawned,
+  SpawnResponse
+} from '../clients/agent-service'
+import { FailedD, LocationServiceErrorD, UnhandledD } from './CommonDecoders'
+import { ComponentIdD } from './ComponentIdDecoder'
 import { ciLiteral, Decoder } from './Decoder'
+import { AkkaLocationD } from './LocationDecoders'
 
 const SpawnedD: Decoder<Spawned> = D.type({
   _type: ciLiteral('Spawned')
@@ -9,6 +20,22 @@ const SpawnedD: Decoder<Spawned> = D.type({
 
 const KilledD: Decoder<Killed> = D.type({
   _type: ciLiteral('Killed')
+})
+
+export const SequenceComponentStatusD: Decoder<SequenceComponentStatus> = D.type({
+  seqCompId: ComponentIdD,
+  sequencerLocation: D.array(AkkaLocationD)
+})
+
+const AgentStatusD: Decoder<AgentStatus> = D.type({
+  agentId: ComponentIdD,
+  seqCompsStatus: D.array(SequenceComponentStatusD)
+})
+
+const AgentStatusSuccessD: Decoder<AgentStatusSuccess> = D.type({
+  _type: ciLiteral('Success'),
+  agentStatus: D.array(AgentStatusD),
+  seqCompsWithoutAgent: D.array(SequenceComponentStatusD)
 })
 
 export const SpawnResponseD: Decoder<SpawnResponse> = D.sum('_type')({
@@ -19,4 +46,10 @@ export const SpawnResponseD: Decoder<SpawnResponse> = D.sum('_type')({
 export const KillResponseD: Decoder<KillResponse> = D.sum('_type')({
   Killed: KilledD,
   Failed: FailedD
+})
+
+export const AgentStatusResponseD: Decoder<AgentStatusResponse> = D.sum('_type')({
+  Unhandled: UnhandledD,
+  LocationServiceError: LocationServiceErrorD,
+  Success: AgentStatusSuccessD
 })