+++ CHANGELOG.md
@@ -4,18 +4,24 @@ ESW-TS is a Javascript/Typescript wrapper for TMT backend services.
 This source code is in github repository: [ESW-TS](https://github.com/tmtsoftware/esw-ts)
 Npm module published for ESW-TS can be found [here](https://www.npmjs.com/package/@tmtsoftware/esw-ts)
 
+## [UPCOMING CHANGES]
+
+* Move of `getAgentStatus` api from SequenceManager service to AgentService
+
 ## [ESW-TS v0.1.0] - 2021-02-01
 
 This is the first minor release of ESW-TS for project stakeholders.
 
 ### Changes
+
 Two main components are delivered as part of ESW-TS <sup>[1](#0-1-0-1)</sup>:
 
 * AAS React Components
 * Typescript Clients for backend services
 
 ### Documentation
-- Reference paradox documentation: https://github.com/tmtsoftware/esw-ts/0.1.0/
+
+* Reference paradox documentation: https://github.com/tmtsoftware/esw-ts/0.1.0/
 
 ### Supporting Releases